pyproject.toml
[tool.poetry]
name = "makedjango"
version = "1.0.0"
description = ""
authors = ["Rust Saiargaliev"]

[tool.poetry.dependencies]
python = "^3.9"
django = "^3.1.7"
django-configurations = "^2.2"
django-environ = "^0.4.5"
django-extensions = "^3.1.1"
django-redis = "^4.12.1"
psycopg2-binary = "^2.8.6"
sentry-sdk = "^0.20.3"

[tool.poetry.dev-dependencies]
bandit = "^1.7.0"
black = "^20.8b1"
django-debug-toolbar = "^3.2"
flake8 = "^3.8.4"
isort = "^5.7.0"
model-bakery = "^1.2.1"
pre_commit = "^2.10.1"
pydocstyle = "^5.1.1"
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
pytest-django = "^4.1.0"
pytest-deadfixtures = "^2.2.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
